+++ src/main/java/ru/topjava/lunchvoter/repository/CrudDishRepository.java
@@ -7,7 +7,7 @@ import ru.topjava.lunchvoter.model.Dish;
 import java.util.List;
 
 @Transactional(readOnly = true)
-public interface CrudDishRepository extends JpaRepository<Dish, Integer> {
+public interface CrudDishRepository extends JpaRepository<Dish, Long> {
     
     List<Dish> getByRestaurantId(Integer restaurantId);
